+++ src/de/dhbw/Microcontroller/Befehle/PIC/BTFSC.java
@@ -16,11 +16,12 @@ public class BTFSC extends Instruction {
 
     @Override
     public void execute(){
-        byte f = (byte) argument1;
+        int f = argument1;
         int b = argument2;
 
-        b = (byte) (f &(1 << (b-1))); //das bit b in f wird verundet um zu prüfen was für ein wert dort steht
+        b = (f &(1 << (b-1))); //das bit b in f wird verundet um zu prüfen was für ein wert dort steht
         if (b==0) {
+            //....
         }
 
         incrementProgramCounter();
